+++ recipes/mod_application_initialization.rb
@@ -19,5 +19,8 @@
 #
 
 include_recipe 'iis'
+install_method = node['iis']['windows_feature_install_method']&.to_sym
 
-windows_feature 'IIS-ApplicationInit'
+windows_feature transform_feature_name(install_method, 'IIS-ApplicationInit') do
+  install_method install_method
+end