+++ recipes/mod_compress_static.rb
@@ -20,4 +20,8 @@
 
 include_recipe 'iis'
 
-windows_feature 'IIS-HttpCompressionStatic'
+install_method = node['iis']['windows_feature_install_method']&.to_sym
+
+windows_feature transform_feature_name(install_method, 'IIS-HttpCompressionStatic') do
+  install_method install_method
+end