+++ recipes/mod_management.rb
@@ -20,7 +20,10 @@
 
 include_recipe 'iis'
 
-windows_feature %w(IIS-ManagementConsole IIS-ManagementService) do
+install_method = node['iis']['windows_feature_install_method']&.to_sym
+
+windows_feature transform_feature_name(install_method, %w(IIS-ManagementConsole IIS-ManagementService)) do
   action :install
   all !IISCookbook::Helper.older_than_windows2012?
+  install_method install_method
 end