+++ resources/install.rb
@@ -18,17 +18,20 @@
 #
 
 include IISCookbook::Helper
+include IISCookbook::WindowsFeatureHelper
 
 property :source, String
 property :additional_components, Array, default: []
+property :install_method, Symbol, required: false, equal_to: [:windows_feature_dism, :windows_feature_powershell, :windows_feature_servermanagercmd], default: :windows_feature_dism
 
 action :install do
-  features_to_install = ['IIS-WebServerRole'].concat new_resource.additional_components
+  features_to_install = transform_feature_name(new_resource.install_method, ['IIS-WebServerRole'].concat(new_resource.additional_components))
 
   windows_feature 'Install IIS and additional components' do
     feature_name features_to_install
     action :install
     all !IISCookbook::Helper.older_than_windows2012?
     source new_resource.source unless new_resource.source.nil?
+    install_method new_resource.install_method
   end
 end