+++ resources/manager.rb
@@ -22,10 +22,12 @@
 property :enable_remote_management, [true, false], default: true
 property :log_directory, String
 property :port, Integer, default: 8172
+property :install_method, Symbol, required: false, equal_to: [:windows_feature_dism, :windows_feature_powershell, :windows_feature_servermanagercmd], default: :windows_feature_dism
 
 action :config do
   iis_install 'Web-Mgmt-Service' do
     additional_components ['IIS-ManagementService']
+    install_method new_resource.install_method
   end
 
   # properties stored in the registry